cilium: nodeid, add reference counting for manager/kvstore upserts
There are two distinct code paths that drive the nodeID life cycle.
The first is the node update/delete path. This path creates a clear
life cycle for the nodeID -- create on update and destroy on delete.
But, there is a second path that is allocated through the ipcache
upsert path. This can happen because of a kvstore update or a sync
with the kernel map from a previous run of the daemon, or through
the watcher code to name a few. These are different from the node
path because they are fundamentally at the IP/pod layer and not at
the node. We also may never see a node update/delete for these because
they could be from another cluster or from restart.

So to manage this IP layer we have code to allocateIDs when the
IP is added by looking up its nodeID. Then if we don't have the
nodeID we allocate and add it. But, we don't have a life cycle
for this path so the nodeID<->IP mapping is never removed.

This can be problematic because once the IPAM believes the IP
is no longer assigned it can readd it on a pod or node update.
Then when nodeID side "sees" this IP it will use the already
allocated ID from a previous iteration. Of course there is no
reason to think this IP belongs to that node so it can result
in incorrect mappings. Further we leak nodeIDs over time. The
concrete result is a missed policy mapping in the xfrm that
results in POLICY BLOCK errors.

Unfortunately, we can't simply delete the nodeID when we see a
delete event on the IP because ew may have multiple IPs from
that node. Sp we introduce reference counting. On ipcache
upsert with a nodeID we inc the refcnt and then when that
IP is released we dec the refcnt. When the refcnt is zero there are
no longer any IPs assigned to it and we free it.

One complication is we can get a nodeUpdate after an IP upsert.
In this case we promote the life cycle ownership over to the
node delete path. This appears to be logical because the node
update/delete path is more explicit.

Signed-off-by: John Fastabend <john.fastabend@gmail.com>